// src/app/models/trendingBlock.model.ts
import { HoverEffectElement } from '../models/hover-effect-element.model';

export class trendingBlock {
    vertical: HoverEffectElement = new HoverEffectElement;
    horizontal1: HoverEffectElement = new HoverEffectElement;
    horizontal2: HoverEffectElement = new HoverEffectElement;
    routerLinkH1: string;
    routerLinkH2: string;
    routerLinkV: string;
}